Remove ldap_str2dn compatibility kludge
This avoids calling the internal, undeclared ldap_bv2dn_x function.
This might be an actual bug because the internal function has since
added an additional argument (which would have to be specified as NULL
in this case).

OpenLDAP 2.2 was released in 2003, so this code should no longer
be needed.